Update to latest uksched and ctx changes
This commit updates the application loading and execution code using
the latest uksched API introduced with PR #564 of the Unikraft core.
Instead of launching the ELF application on the same stack, the ELF
application is executed in a new thread context. The elfloader main
context is kept, so that we can later wait and handle an application
exit. This commit also tries to produce minimal builds: For instance,
the dependency to `lib/posix-process` is optional but recommended.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Razvan Deaconescu <razvand@unikraft.io>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #6