Activate userland TLS support
Because ELF applications that are compiled for Linux assume exclusive
use of the TLS register (see `arch_prctl.c`) we have to enable support
for userland TLS at `lib/syscall_shim`. The setting enables a TLS switch
between the domain of the ELF application and the Unikraft kernel.
On each binary system call, the Unikraft TLS is active for handling
the system call. On its return to the ELF application domain (userland)
the application's TLS register value is restored.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>