Add support for loading ELF images from the root file system
This change introduces support for InitRD file system. Previously, initrd used to
point to the Linux ELF image that was supposed to be loaded, meaning that we could
not populate the RamFS with a cpio archive and therefore we could not run
dynamically-linked Linux ELFs (because `.so` dependencies could not be loaded).

Now, if `CONFIG_LIBVFSCORE_ROOTFS_INITRD` is set, app-elfloader will load the
ELF image and its dependencies (if any) from the file system.

Also, this commit removes the custom app name build configuration. It seems that
the normal configuration already supports this use-case (running apps with
multiple "personalities"). Moreover, the option seems to be broken.

Signed-off-by: Cosmin Vancea <csvancea@gmail.com>