elf_ctx.c: Forward environmental variables
This commit forwards the environment variables that are managed by
`lib/posix-environ` to the ELF application.
The implementation assumes that only the ELF program is executed and that
environment variables are not modified/released by `lib/posix-environ`
while the ELF program runs. This allows handing-over a reference of
each environment variable to the application instead copying them.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>
Reviewed-by: Marc Rittinghaus <marc.rittinghaus@unikraft.io>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #17