elf_load.c: Restructure
This commit restructures `elf_load_img()` as preparation for VFS support.
Main motivation is code re-use for all ways of loading ELF applications.
For this purpose, the function is split into API functions for "checking",
"loading", and "protecting". Error codes are returned as negative value.
The use of `errno` is avoided.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>