brk: Return address from previous brk context on follow-up request
Currently, if there is a follow-up brk request, an error message is
shown and an ENOMEM error is returned.

"Cannot handle multiple user space heaps: Not implemented!"

For now, update the brk logic so that it returns the address from the
previous brk context, when outside of the brk address range.

Fix typo in the comment for the static variables, to mention "heap"
instead of "stack".

Signed-off-by: Andra Paraschiv <andra@unikraft.io>
Reviewed-by: Simon Kuenzer <simon@unikraft.io>
Approved-by: Simon Kuenzer <simon@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #15