elf_load.c: Use `lib/ukvmem` for page protection
Instead of using the low-level paging API, this commit lets elfloader to
interface to `lib/ukvmem` for setting page protection bits. The reason is
that `lib/ukvmem` is keeping record of page mappings and should not be
by-passed so that faults can be handled appropriately (like on-demand
paging). It is built on top of the paging API.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>