elf_load.c, elf_ctx.c: Native dynamic loader support (.interp)
This commit implements native support for program interpreters which are
typically dynamic loaders. Such programs are announced with a path in the
`.interp` section of the actual ELF application. The interpreter is side-
loaded with the application and its entry point needs to be entered instead
of the application ones. The interpreter/dynamic loader is then able to map
depending shared libraries into memory and link them to the application. It
is also the interpreter/dynamic loader that will afterwards enter the
applications entry point after it finished its tasks.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>