elf_ctx.c: Remove vDSO-related auxiliary vector entries
The auxiliary vector fields `AT_SYSINFO` and `AT_SYSINFO_EHDR` are used to
announce a kernel-mapped vDSO (virtual dynamic shared object) to the
application in order to provide system call alternatives that do not
perform a user/kernel mode switch and typically faster to handle.
Because we do currently not support this mechanism, we announced a vDSO by
pointing to the zero page. However, this may lead for some applications to
lookup vDSO symbols at the zero page instead of ignoring it. The correct
behavior is not to announce a vDSO at all.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>
Reviewed-by: Marc Rittinghaus <marc.rittinghaus@unikraft.io>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #17