main: Random seed from `lib/ukswrand`
The ELF auxiliary vector provides fields for a random number seed that the
application can use for initializing an own random number generator. This
commit initializes this seed with random numbers retrieved from
`lib/ukswrand` if selected as part of the build. Because not having proper
initialized random number generators can be a serious security concern,
we print a warning whenever we fall back to a compiled-in constant.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>