Introduce build and run scripts
Introduce quick scripts to build both with Make and KraftKit and to run
both with KraftKit, QEMU and Firecracker.

The commit effectively consists of a single script: `generate.py`.
`generate.py` is be used to generate the actual scripts in the
`scripts/build/` and `scripts/run/` directories. This will improve
maintenance (only required to happen for the `generate.py` script),
while still keeping things usable by providing the basic build and run
scripts.

The generator script provides maintainability (not repeating
the same commited information) and also usability by providing the
actual scripts to use. Users can then quickly check the scripts and
update them according to their needs.

The `scripts/run.yaml` file contains runtime configuration for
applications to be run: memory, command to be used, filesystem type.

The generator script should be common to all applications. So it is
meant to be moved in a common repository. For now, keept it part of this
repository, until the app catalog repository becomes available.

Signed-off-by: Razvan Deaconescu <razvand@unikraft.io>