`mmap` each ELF segment if `CONFIG_LIBPOSIXMMAP` is enabled
Add an additional way of loading an ELF depending on whether
CONFIG_LIBPOSIX_MMAP is used or not. If it is enabled, then the
ELF loader will `mmap` the file instead of `pread` it. Otherwise,
the previous method of `pread`ing file contents is used.

The first address of the executable will be the first result of the
first time call to `mmap` on the first `PT_LOAD` and everything else
will be contiguously mapped by using the `MAP_FIXED` flag.

Some special things to be noted when it comes to misaligned end
addresses of ELF segments. Suppose we have `vastart` and `vaend`,
corresponding to the start address of the ELF segment and the end
address respectively and consider memsz as the runtime memory size
occupied by the loaded segment:
if vaend > vastart + memsz
	if vaend == PAGE_ALIGN_UP(vastart + memsz)
		the remaining bytes from (vastart + memsz) to vaend
		have already been mapped, so simply memset with 0
	else /* vaend > PAGE_ALIGN_UP(vastart + memsz) */
		this segment contains a section with NOBITS, such as
		.bss, which can be fairly large. Therefore:
		/* memset what is already mapped */
		memset(vastart + memsz, 0,
		       PAGE_ALIGN_UP(vastart + memsz) - vastart);
		/* anonymously map what is left, as memsetting large
		 * sections such as .bss is very costly.
		 */
		mmap(PAGE_ALIGN_UP(vastart + memsz),
		     vaend - PAGE_ALIGN_UP(vastart + memsz),
		     ..., MAP_ANONYMOUS, ...);

NOTE: The first segment must be mmap'd with the length of the entire
file in memory and then the surplus is to be `munmapped`. This way we
ensure we will always have a contiguous mapping big enough ready.
Although future `mmap`'s with `MAP_FIXED` should normally overwrite
this surplus, ridding us of the need to explicitly `munmap` it, in
this current state of Unikraft (0.14.1) there is a bug in `ukvmem`
that may not handle overlapping VMA's correctly.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>