Makefile.uk: Embed Makefile.rules
`Makefile.rules` is a reserved build system file that is intended to add further
build system features, like integrating additional rules for compiled
languages. Such a `Makefile.rules` file is normally automatically included and
loaded before library context is created. It is not intended to be used for
compiling files for a library or application.
This commit moves the content from `Makefile.rules` to `Makefile.uk`. Compiling
the VDSO should be part of building `appelfloader`.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>