Simplify the mapping of the first ELF segment
In order to make sure that subsequent `mmap` calls with `MAP_FIXED`
give us contiguous mappings we must ensure that the first
`mmap` with argument `NULL` for address will yield us the address
of a region large enough for the subsequent mappings.

Before, we would do an initial large `mmap` (the length also contains
the alignment bias) of the file starting the offset of the first ELF
segment and then, depending on whether the received start address has
the desired alignment or not, we do another `mmap` of this first segment
with `MAP_FIXED` over this initial dummy `mmap` to overwrite its VMA and
then do a `munmap` over the surplus area, beyond what we need for the
actual first ELF segment.

We can simplify this by making this initial dummy `mmap` an anonymous
mapping (this does not really make any difference, it's just that it does
not need to be a file mapping that would require `mmap` to do some
additional things setup-wise) of the same size, keeping this initially
received address and then doing a `munmap` of this whole area.
Afterwards, we do the segment mapping as normal, starting from this
address. This way, `ukvmem` does not have to deal with figuring out
VMA splitting which, at the moment (Unikraft 0.15), is not fully stable.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>