Correctly set `Program Header` offset used for aux vector's `AT_PHDR`
The Program Header offset specified in the ELF Header represents the
in-file offset. If this happens to also be part of a PT_LOAD segment,
this could mean trouble: in-file offsets != in-memory offset.
Although they are almost always the same, it is not uncommon to meet
ELF's whose segment in-file offsets differ from those in memory at
runtime. Thus, this means that we can't actually use the in-file
offset of the Program Headers.

Fix this by figuring out the PT_LOAD ELF segment the Program Headers
are part of during `elf_load_parse`: if the in-file offset of the
Program Headers is contained within the in-file contents of that
segment, then we found it. Consequently, obtain the in-memory offset
of the Program Headers by finding the in-file offset of the Program
Headers relative to the start of the segment they are part of
(ehdr->e_phoff - phdr->p_offset) and adding to that the runtime
base address of the ELF segment ( + phdr->p_paddr).

Finally, compute `AT_PHDR` based on this new offset and the virtual
address base (elf_prog->vabase), not the start of the actual ELF in
its mapped area (elf_prog->vastart) as they could differ.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>