Set `mmap` permission flags only the first time we `mmap` a segment
Previously, for simplicity, the ELF loader would initially `mmap` all
segments with RWX permissions only to then let `elf_load_ptprotect`
apply proper mappings. However, this turned out to be unnecessarily
suboptimal as this would make `lib/ukvmem` unmap, unlink and free the
previously mentioned VMA in favor of creating and adding a new VMA
that is equivalent to the deleted one except for the permissions.

Thus, avoid this and use proper permissions the first time we `mmap`
an ELF segment and no longer execute `elf_load_ptprotect` afterwards
if `CONFIG_LIBPOSIX_MMAP` is enabled.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
Reviewed-by: Michalis Pappas <michalis@unikraft.io>
GitHub-Closes: #64